+++ include/ice/manage/dec/IDecoder.hpp
@@ -17,9 +17,9 @@ class IDecoder {
     // 获取总帧数量接口
     virtual size_t num_frames() const = 0;
 
-    // 转移数据到缓冲区的接口
-    virtual size_t decode(float** buffer, uint16_t num_channels,
-                          size_t start_frame, size_t frame_count) = 0;
+    // 转移数据到缓冲区的接口(读取数量使用浮点数)
+    virtual double decode(float** buffer, uint16_t num_channels,
+                          double start_frame, double frame_count) = 0;
 };
 
 }  // namespace ice
