+++ src/ice/manage/dec/StreamingDecoder.cpp
@@ -14,8 +14,8 @@ StreamingDecoder::StreamingDecoder(std::string_view path,
                                    std::shared_ptr<IDecoderFactory> factory) {}
 
 // 解码数据到缓冲区的接口
-size_t StreamingDecoder::decode(float** buffer, uint16_t num_channels,
-                                size_t start_frame, size_t frame_count) {
+double StreamingDecoder::decode(float** buffer, uint16_t num_channels,
+                                double start_frame, double frame_count) {
     // TODO(xiang 2025-07-11): 实现流式解码
     return 0;
 }
